refactor statemap as subcomponent of state view
